+++ gimmemotifs/scanner/base.py
@@ -32,8 +32,8 @@ from gimmemotifs.scanner.utils import (
 from gimmemotifs.utils import as_fasta
 
 logger = logging.getLogger("gimme.scanner")
-config = MotifConfig()
 FPR = 0.01
+THRESHOLD = 0.95
 LOCK = Lock()
 
 
@@ -42,17 +42,24 @@ class Scanner(object):
     scan sequences with motifs
     """
 
-    def __init__(self, ncpus=None):
+    genome = None
+    background = None
+    background_hash = None
+    meanstd = {}
+    gc_bins = [(0, 1)]
+    fpr = None
+    motifs = None
+    motif_ids = []
+    _threshold = None
+    _seed = None
+
+    def __init__(self, ncpus=None, random_state=None, progress=None):
         self.config = MotifConfig()
-        self._threshold = None
-        self.genome = None
-        self.background = None
-        self.meanstd = {}
-        self.gc_bins = [(0, 1)]
-        self.fpr = None
+        self.random_state = random_state
+        self.progress = progress
 
         if ncpus is None:
-            ncpus = MotifConfig().get_default_params()["ncpus"]
+            ncpus = self.config.get_default_params()["ncpus"]
         self.ncpus = int(ncpus)
 
         if self.ncpus > 1:
@@ -65,6 +72,30 @@ class Scanner(object):
             self.pool.join()
         except (ValueError, AttributeError):
             pass
+
+    @property
+    def random_state(self):
+        return self._random_state
+
+    @random_state.setter
+    def random_state(self, random_state):
+        self._random_state = random_state
+        if random_state is not None:
+            # can be used to set random.seed()
+            self._seed = random_state.get_state()[1][0]
+
+    @property
+    def progress(self):
+        """Whether to show progress bars for slow operations."""
+        return self._progress
+
+    @progress.setter
+    def progress(self, progress):
+        self._progress = progress
+        if progress is None:
+            self._disable = None
+        else:
+            self._disable = not progress
 
     def set_motifs(self, motifs):
         try:
@@ -91,46 +122,28 @@ class Scanner(object):
         for (motif, _), scores in zip(scan_motifs, np.array(table).transpose()):
             yield motif, np.mean(scores), np.std(scores)  # cutoff
 
-    def _threshold_from_seqs(self, motifs, seqs, fpr):
+    def _threshold_from_seqs(self, motifs, seqs):
         scan_motifs = [(m, m.min_score) for m in motifs]
-        table = []
         seq_gc_bins = [self.get_seq_bin(seq) for seq in seqs]
+
+        # progress bar
+        pbar = tqdm(
+            desc="Determining FPR-based threshold",
+            unit=" sequences",
+            total=len(seqs),
+            disable=self._disable,  # can be silenced
+        )
+
+        table = []
         for gc_bin, result in zip(
             seq_gc_bins, self._scan_sequences_with_motif(scan_motifs, seqs, 1, True)
         ):
             table.append([gc_bin] + [row[0][0] for row in result])
+            pbar.update(1)
 
         df = pd.DataFrame(table, columns=["gc_bin"] + [m.id for m in motifs])
+        df.set_index("gc_bin", inplace=True)
         return df
-
-    # the functions below are conceptually simpler (and faster?) but don't work due to
-    # multiprocessing.Pool not being able to be used from another Pool.
-    # Future improvement to fix this.
-    # def _meanstd_from_seqs(self, motifs, seqs):
-    #     # We need a new scanner instance, as this may be a subset of motifs
-    #     s = Scanner()
-    #     s.set_genome(self.genome)
-    #     s.set_motifs(motifs)
-    #     table = np.array(list(s.best_score(seqs, scan_rc=True, zscore=False, gc=False, progress=False)))
-
-    #     for motif, scores in zip(motifs, table.transpose()):
-    #         yield motif, np.mean(scores), np.std(scores)  # cutoff
-
-    # def _threshold_from_seqs(self, motifs, seqs, fpr):
-    #     # We need a new scanner instance, as this may be a subset of motifs
-    #     s = Scanner()
-    #     s.set_genome(self.genome)
-    #     s.set_motifs(motifs)
-    #     table = []
-    #     seq_gc_bins = [self.get_seq_bin(seq) for seq in seqs]
-    #     for gc_bin, row in zip(
-    #         seq_gc_bins, list(s.best_score(seqs, scan_rc=True, zscore=False, gc=False, progress=False))
-    #     ):
-    #         print(row)
-    #         table.append([gc_bin] + list(row)
-
-    #     df = pd.DataFrame(table, columns=["gc_bin"] + [m.id for m in motifs])
-    #     return df
 
     def set_meanstd(self, gc=False):
         if not self.background:
@@ -140,35 +153,47 @@ class Scanner(object):
         seqs = self.background.seqs
         if gc:
             seq_bins = [s.split(" ")[-1] for s in self.background.ids]
+            bins = sorted(set(seq_bins))
         else:
             seq_bins = ["0.00-1.00"] * len(seqs)
-        if gc:
-            bins = list(set(seq_bins))
-        else:
             bins = ["0.00-1.00"]
 
         motifs = read_motifs(self.motifs)
         LOCK.acquire()
-
         try:
             with Cache(CACHE_DIR) as cache:
-                scan_motifs = []
+
+                # for each bin, load the meanstd of each motif or mark it as missing
+                scan_gc_bins = {}
                 for bin in bins:
                     if bin not in self.meanstd:
                         self.meanstd[bin] = {}
-                    bin_seqs = [s for s, b in zip(seqs, seq_bins) if b == bin]
 
                     for motif in motifs:
                         k = "e{}|{}|{}".format(motif.hash, self.background_hash, bin)
-
                         results = cache.get(k)
                         if results is None:
-                            scan_motifs.append(motif)
+                            # missing values
+                            scan_gc_bins.setdefault(bin, []).append(motif)
                         else:
                             self.meanstd[bin][motif.id] = results
 
-                    if len(scan_motifs) > 0:
-                        logger.debug("Determining mean and stddev for motifs.")
+                # generate missing values for each motif per gc_bin
+                total_scans = sum(len(v) for v in scan_gc_bins.values())
+                if total_scans > 0:
+                    pbar = tqdm(
+                        desc="Determining mean and stddev for motifs",
+                        unit=" motifs",
+                        total=total_scans,
+                        disable=self._disable,  # can be silenced
+                    )
+
+                    for bin, scan_motifs in scan_gc_bins.items():
+                        bin_seqs = [s for s, b in zip(seqs, seq_bins) if b == bin]
+                        if len(bin_seqs) == 0:
+                            # no background sequences with this GC%
+                            pbar.update(len(scan_motifs))
+                            continue
                         for motif, mean, std in self._meanstd_from_seqs(
                             scan_motifs, bin_seqs
                         ):
@@ -177,6 +202,7 @@ class Scanner(object):
                             )
                             cache.set(k, [mean, std])
                             self.meanstd[bin][motif.id] = mean, std
+                            pbar.update(1)
 
                 # Prevent std of 0
                 # This should only happen in testing
@@ -194,6 +220,7 @@ class Scanner(object):
                             mean = self.meanstd[gcbin][motif.id][0]
                             cache.set(k, [mean, std])
                             self.meanstd[gcbin][motif.id] = mean, std
+
         except sqlite3.DatabaseError:
             print_cluster_error_message()
             sys.exit(1)
@@ -207,6 +234,13 @@ class Scanner(object):
                     bstr = f"{b[0]:.2f}-{b[1]:.2f}"
                     if bstr in self.meanstd:
                         valid_bins.append(((b[0] + b[1]) / 2, bstr))
+
+                if len(valid_bins) == 0:
+                    logger.error(
+                        f"No sequences found fo gc bin '{gc_bin}'. "
+                        "This likely occurred because you set gc=True with a background file"
+                    )
+                    sys.exit()
 
                 v = float(gc_bin.split("-")[1])
                 _, bstr = sorted(valid_bins, key=lambda x: abs(x[0] - v))[0]
@@ -242,9 +276,10 @@ class Scanner(object):
 
         if gc_bins is None:
             if gc:
-                gc_bins = [(0.0, 0.2), (0.8, 1)]
+                gc_bins = [(0.0, 0.2), (0.8, 1.0)]
                 for b in np.arange(0.2, 0.799, 0.05):
-                    gc_bins.append((b, b + 0.05))
+                    gc_bins.append((round(b, 2), round(b + 0.05, 2)))
+                gc_bins = sorted(gc_bins)
             else:
                 gc_bins = [(0, 1)]
         if nseq is None:
@@ -286,11 +321,11 @@ class Scanner(object):
                         with NamedTemporaryFile() as tmp:
                             logger.info(f"using {nseq} sequences")
                             gc_bin_bedfile(
-                                tmp.name, genome, number=nseq, length=size, bins=gc_bins
+                                tmp.name, genome, nseq, size, gc_bins, self.random_state
                             )
                             fa = as_fasta(tmp.name, genome=genome)
                     else:
-                        fa = RandomGenomicFasta(genome, size, nseq)
+                        fa = RandomGenomicFasta(genome, size, nseq, self._seed)
                     cache.set(self.background_hash, (fa, gc_bins))
         except sqlite3.DatabaseError:
             print_cluster_error_message()
@@ -327,10 +362,10 @@ class Scanner(object):
 
         if threshold is None and fpr is None:
             if self.genome:
-                fpr = 0.01
+                fpr = FPR
                 logger.info(f"Using default FPR of {fpr}")
             else:
-                threshold = 0.95
+                threshold = THRESHOLD
                 logger.info(
                     f"Genome not specified, using default threshold of {threshold}."
                 )
@@ -388,9 +423,7 @@ class Scanner(object):
 
                 if len(scan_motifs) > 0:
                     logger.debug("determining FPR-based threshold")
-                    df = self._threshold_from_seqs(scan_motifs, seqs, fpr).set_index(
-                        "gc_bin"
-                    )
+                    df = self._threshold_from_seqs(scan_motifs, seqs)
                     if self._threshold is None:
                         self._threshold = df
                     else:
@@ -407,17 +440,14 @@ class Scanner(object):
             print_cluster_error_message()
             sys.exit(1)
         LOCK.release()
-        self.threshold_str = "{}_{}_{}_{}".format(
-            fpr, threshold, self.background_hash, ",".join(sorted(gc_bins))
-        )
 
-    def set_genome(self, genome, genomes_dir=None):
+    def set_genome(self, genome=None, genomes_dir=None):
         """
         Set the genome to converting regions to sequences
 
         Parameters
         ----------
-        genome : str
+        genome : str, optional
             Path to the genome fasta or
             the genome name as found in the genomepy genomes directory.
 
@@ -425,58 +455,50 @@ class Scanner(object):
             Path to the genomepy genomes directory.
             Taken from the genomepy config if unspecified.
         """
-        if not genome:
-            return
-
-        # raises error if checks fail
-        genome = Genome(genome, genomes_dir, rebuild=False).filename
+        if genome is not None:
+            # raises error if checks fail
+            genome = Genome(genome, genomes_dir, rebuild=False).filename
 
         self.genome = genome
 
-    def count(self, seqs, nreport=100, scan_rc=True, progress=True):
+    def count(self, seqs, nreport=100, scan_rc=True):
         """
         count the number of matches above the cutoff
         returns an iterator of lists containing integer counts
         """
-        for matches in self.scan(seqs, nreport, scan_rc, progress=progress):
+        for matches in self.scan(seqs, nreport, scan_rc):
             counts = [len(m) for m in matches]
             yield counts
 
-    def total_count(self, seqs, nreport=100, scan_rc=True, progress=True):
+    def total_count(self, seqs, nreport=100, scan_rc=True):
         """
         count the number of matches above the cutoff
         returns an iterator of lists containing integer counts
         """
 
-        count_table = [
-            counts for counts in self.count(seqs, nreport, scan_rc, progress=progress)
-        ]
+        count_table = [counts for counts in self.count(seqs, nreport, scan_rc)]
         return np.sum(np.array(count_table), 0)
 
-    def best_score(self, seqs, scan_rc=True, zscore=False, gc=False, progress=True):
+    def best_score(self, seqs, scan_rc=True, zscore=False, gc=False):
         """
         give the score of the best match of each motif in each sequence
         returns an iterator of lists containing floats
         """
         self.set_threshold(threshold=0.0, gc=gc)
-        for matches in self.scan(
-            seqs, 1, scan_rc, zscore=zscore, gc=gc, progress=progress
-        ):
+        for matches in self.scan(seqs, 1, scan_rc, zscore=zscore, gc=gc):
             scores = np.array(
                 [sorted(m, key=lambda x: x[0])[0][0] for m in matches if len(m) > 0]
             )
             yield scores
 
-    def best_match(self, seqs, scan_rc=True, zscore=False, gc=False, progress=True):
+    def best_match(self, seqs, scan_rc=True, zscore=False, gc=False):
         """
         give the best match of each motif in each sequence
         returns an iterator of nested lists containing tuples:
         (score, position, strand)
         """
         self.set_threshold(threshold=0.0)
-        for matches in self.scan(
-            seqs, 1, scan_rc, zscore=zscore, gc=gc, progress=progress
-        ):
+        for matches in self.scan(seqs, 1, scan_rc, zscore=zscore, gc=gc):
             yield [m[0] for m in matches]
 
     def get_seq_bin(self, seq):
@@ -488,7 +510,7 @@ class Scanner(object):
         if gc == 0:
             gc = 0.01
         for b_start, b_end in self.gc_bins:
-            if gc > round(b_start, 2) and gc <= round(b_end, 2):
+            if round(b_start, 2) < gc <= round(b_end, 2):
                 return f"{b_start:.2f}-{b_end:.2f}"
 
         logger.error(f"Error determining seq: {seq}, bins: {str(self.gc_bins)}")
@@ -519,9 +541,7 @@ class Scanner(object):
             self.meanstd[gc_bin] = self.meanstd[bstr]
         return self.meanstd[gc_bin][motif]
 
-    def scan(
-        self, seqs, nreport=100, scan_rc=True, zscore=False, gc=False, progress=True
-    ):
+    def scan(self, seqs, nreport=100, scan_rc=True, zscore=False, gc=False):
         """
         Scan a set of regions or sequences.
         """
@@ -536,20 +556,19 @@ class Scanner(object):
 
         # progress bar
         pbar = tqdm(
-            desc="scanning",
+            desc="Scanning",
             unit=" sequences",
             total=len(seqs),
-            disable=not progress,  # can be silenced
+            disable=self._disable,  # can be silenced
         )
 
         batch_size = 50000
-        logger.debug("Scanning")
         for batch_idx in range(0, len(seqs), batch_size):
             it = self._scan_sequences(
                 seqs.seqs[batch_idx : batch_idx + batch_size],
                 nreport,
                 scan_rc,
-                zscore=zscore,
+                zscore,
             )
             for result in it:
                 yield result
@@ -557,14 +576,14 @@ class Scanner(object):
         pbar.close()
 
     def get_gc_thresholds(self, seqs, motifs=None, zscore=False):
-        MAX_SEQS = 20000
-
         # Simple case, only one threshold
         if np.all(self.threshold.nunique(axis=0) == 1):
             return self.threshold.iloc[0].to_dict()
 
-        if len(seqs) > MAX_SEQS:
-            seqs = np.random.choice(seqs, size=MAX_SEQS)
+        max_seqs = 20000
+        if len(seqs) > max_seqs:
+            random = np.random if self.random_state is None else self.random_state
+            seqs = random.choice(seqs, size=max_seqs)
 
         if motifs is None:
             motifs = read_motifs(self.motifs)
@@ -581,7 +600,7 @@ class Scanner(object):
                 columns=_threshold.columns,
             )
 
-        nseqs = int(MAX_SEQS / np.sum(list(gc_bin_count.values())))
+        nseqs = int(max_seqs / np.sum(list(gc_bin_count.values())))
         t = {}
         maxt = pd.Series([m.max_score for m in motifs], index=_threshold.columns)
         # We do this in a loop as the DataFrame will get too big to fit in memory
@@ -591,7 +610,7 @@ class Scanner(object):
         idx = np.hstack(
             [
                 _threshold[_threshold[_threshold.columns[0]] == gc_bin]
-                .sample(nseqs * count, replace=True, random_state=42)
+                .sample(nseqs * count, replace=True, random_state=self.random_state)
                 .index.values
                 for gc_bin, count in gc_bin_count.items()
             ]