+++ CHANGELOG.md
@@ -7,12 +7,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- `Scanner.set_genome` now accepts the optional argument `genomes_dir`
 - `gimme diff` (`diff_plot()` to be exact) will now print to stdout, like all other functions
 - now using a logger instead of print/sys.stderr.write in many more places
 - string formatting now (mostly) done with f-strings
 - refactored Fasta class
-- removed old python2 code (MOODS & import shenanigans)
+- removed old python2 code (scanning with MOODS & import shenanigans)
+- `Scanner` now accepts a `np.random.RandomState` and progress on init.
+  - `progress=None` (the default) should print progress bars to the command line only, not to file.
+- `Scanner.set_genome` now accepts the optional argument `genomes_dir`
 - split `scanner.py` into 3 submodules:
   - `scanner/__init__.py` with the exported functions
   - `scanner/base.py` with the Scanner class